#54 - code review - move symbolic alleles into constants rather than magic strings everywhere
