+++ upload/vcf/bulk_genotype_vcf_processor.py
@@ -10,7 +10,8 @@ from django.conf import settings
 
 from library.django_utils import thread_safe_unique_together_get_or_create
 from library.django_utils.django_file_utils import get_import_processing_filename
-from library.genomics.vcf_utils import VCFConstant, vcf_get_ref_alt_end
+from library.genomics.vcf_enums import VCFConstant
+from library.genomics.vcf_utils import vcf_get_ref_alt_end
 from library.git import Git
 from library.utils import double_quote
 from library.utils.database_utils import postgres_arrays
